+++ pkg/target/target.go
@@ -69,9 +69,10 @@ func Glob(dst string, globs ...string) (bool, error) {
 
 // Dir reports whether any of the sources have been modified more recently
 // than the destination. If a source or destination is a directory, this
-// function returns true if a source has any file that has been modified more
-// recently than the most recently modified file in dst. If the destination
-// file doesn't exist, it always returns true and nil.  It's an error if any
+// function returns true if a source has any file (excluding the directory
+// itself) that has been modified more recently than the most recently modified
+// file in dst (also excluding the directory itself). If the destination
+// file doesn't exist, it always returns true and nil. It's an error if any
 // of the sources don't exist.
 //
 // Dir respects the global ignorelist populated by AddIgnorePattern and